src/main/java/com/leetcode/TwoSum.java
package com.leetcode;

import java.util.HashMap;
import java.util.Map;

/**
 * Solution for TwoSum problem.
 * https://leetcode.com/problems/two-sum/
 *
 * @author Vasyl Grygoryev (aka Basil Gregory)
 *         created on 3/9/19
 */
public class TwoSum {

    public int[] twoSum(int[] nums, int target) {
        Map<Integer, Integer> map = new HashMap<>();

        for (int i = 0; i < nums.length; i++) {
            int diff = target - nums[i];
            if (map.containsKey(diff)) {
                return new int[] {map.get(diff), i};
            }
            map.put(nums[i], i);
        }

        return null;
    }

}
